refactor: 설정 컴포넌트 수정
